+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## UNRELEASE
 
 - **Fix** : Fixed payment product api store id
+- **Fix** : Fixed default env stage value in config.xml for payment methods
+- **Fix** : Improved efficiency of PayPal config provider service
 
 ## 1.27.5